Revert "Merge branch 'master' into dr-feature"
This reverts commit 761d3554ec88298e3e7595efb4c427a428d020ab, reversing
changes made to 726eb4a8ab135ae88ad04b16f4fee12a8c13c523.

+++ tests/com/picsauditing/PICS/flags/MultiYearValueCalculatorTest.java
@@ -128,8 +128,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(11);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(1, result.intValue());
+		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(1, result);
 	}
 	
 	@Test
@@ -139,8 +139,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(8);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(0, result.intValue());
+		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(0, result);
 	}
 	
 	@Test
@@ -150,8 +150,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(10);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(0, result.intValue());
+		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(0, result);
 	}
 
 	@Test
